requirements.txt
berserk
matplotlib
datetime
dateutils
prophet